# server/common/api.yml
openapi: 3.0.1
info:
  title: eric-express-server
  description: Eric&#39;s demo express server using TypeScript
  version: 1.0.0
servers:
  - url: /api/v1
tags:
  - name: Examples
    description: Simple example endpoints
  - name: Specification
    description: The swagger API specification
paths:
  /examples:
    get:
      tags:
        - Examples
      description: Fetch all examples
      responses:
        200:
          description: Returns all examples
          content: {}
    post:
      tags:
        - Examples
      description: Create a new example
      requestBody:
        description: an example
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExampleRequestNoID'
        required: true
      responses:
        201:
          description: Returns example created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
  /examples/{id}:
    get:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          description: The id of the example to retrieve
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the example with the specified id
          content: {}
        404:
          description: Example not found
          content: {}
    delete:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          description: The id of the example to delete
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the deleted example with the specified id
          content: {}
        404:
          description: Example not found
          content: {}
    put:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          description: The id of the example to create/update
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the example being updated
          content: {}
        201:
          description: Return the example being created
          content: {}
    patch:
      tags:
        - Examples
      parameters:
        - name: id
          in: path
          description: The id of the example to update
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the example being updated
          content: {}
        404:
          description: Example not found
          content: {}
  /test:
    get:
      tags:
        - test
      description: Fetch all test
      responses:
        200:
          description: Returns all examples
          content: {}
    post:
      tags:
        - test
      description: Create a new test
      responses:
        201:
          description: Returns test created
          content: {}
  /products:
    get:
      tags:
        - product
      description: Fetch all products
      responses:
        200:
          description: Returns all products
          content: {}
  /spec:
    get:
      tags:
        - Specification
      responses:
        200:
          description: Return the API specification
          content: {}
components:
  schemas:
    ExampleRequestNoID:
      title: example
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          example: 10
        name:
          type: string
          example: no_stress
    ExampleResponse:
      title: example
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          example: 10
        name:
          type: string
          example: no_stress